Update Chinese presigned put example (#258)
The Chinese version of the Upload Files Using Pre-signed URLs document
has been updated with the changes from #256

It should be noted that the comments added to the code since have been
kept in English to prevent any inaccuracies and misunderstanding. These
comments can be revisited in the next iteration of translating.