Updating base URLs for crawler and intersphinx links (#561)
Algolia crawler requires the baseURLs be bare domains, otherwise a conflict arises between the loaded page and the defined canonical URL in each page. This causes the Algolia crawler to reject the page and not find anything to index in the docs.

Also corrects Intersphinx errors at build time resulting from not having the correct paths to linux and kubernetes docs.