Add clickable buttons to navigate between images in the carousel (#714)
Preview:
<img width="992" alt="Screenshot 2023-01-25 at 20 03 00"
src="https://user-images.githubusercontent.com/13393018/214613411-28d2b9b9-2031-4d6a-8aac-3ddeea763199.png">

Limitations:
- Arrows are only present on the carousel and not in the Lightbox
preview. In the future, I'll investigate the options for including
arrows on the Light box. For this, we might need to replace the existing
Lightbox plugin with a modern one.
- Loop navigation is not available; instead, arrows will only be shown
when they correspond to the first and last slides.