Adding staging logic (#759)
# Summary

The actual logic is pretty straightforward - run `mc cp -r` to
recursively copy the contents of the correct directory out to a MinIO
alias and bucket.

For now, the alias and bucket are hardcoded. We could make that more
variable, but this is easier.

This will require you to install `mc` locally and ensure you have the
`docs-staging` access setup . check slack for the details - if `mc ls
docs-staging/staging` shows you a list of already staged docs, you're in
good shape.