Remote tier needs DeleteObjectVersion if remote bucket is versioned (#1089)
If the remote tier bucket is versioned, MinIO requires additional
permissions to successfully transition objects. But it is not
recommended, as each object version becomes a distinct object on the
remote tier.

This is not new, it was omitted from the docs. Mention, but discourage.

Staged:

http://192.241.195.202:9000/staging/DOCS-1016/linux/reference/minio-mc/mc-ilm-tier.html#transition-permissions

Fixes https://github.com/minio/docs/issues/1016

---------

Co-authored-by: Daryl White <53910321+djwfyi@users.noreply.github.com>