DOCS-1190: RELEASE.2024-04-18T19-09-19Z (#1215)
Partially addresses #1190 

No staged build unless someone requests it.

# Summary

After some thinking, we've leaned towards just making a note of the
upgrade step and calling it done. We do not typically make extended
discussion of backward-breaking changes at this time, and have no where
to keep that sort of guidance.

Maybe in the future or as part of a larger re-arch we can consider where
we would best keep that kind of guidance.

As far as the rest:

- We are not covering s3zip today, so that's a no-op. We should leave
the ticket open in the event we have bandwidth to document the feature
later
- No changes to Console that need documentation at this time
- Added a quick callout on the warning behavior to the Versioning page
- Metrics v3 is also still In progress, come back to this once the
project is fully baked.